+++ src/Controller/BO/SocieteController.php
@@ -6,8 +6,8 @@ use App\DTO\InitSociete;
 use App\Entity\Societe;
 use App\Entity\SocieteUser;
 use App\Form\InitSocieteType;
+use App\Form\OnboardingNotificationEveryType;
 use App\Form\UserEmailType;
-use App\License\DTO\License;
 use App\License\Factory\OffreStarterLicenseFactory;
 use App\License\LicenseService;
 use App\LicenseGeneration\Exception\EncryptionKeysException;
@@ -16,11 +16,9 @@ use App\LicenseGeneration\LicenseGeneration;
 use App\Repository\SocieteRepository;
 use App\Security\Role\RoleSociete;
 use App\File\FileResponseFactory;
-use App\Form\ParameterType;
-use App\Repository\ParameterRepository;
 use App\Service\Invitator;
+use App\SocieteProduct\Product\ProductPrivileges;
 use Doctrine\ORM\EntityManagerInterface;
-use League\Flysystem\FilesystemInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -28,34 +26,16 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\HttpKernel\Exception\HttpException;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Contracts\Translation\TranslatorInterface;
 
 class SocieteController extends AbstractController
 {
     /**
      * @Route("/societes", name="app_bo_societes")
      */
-    public function societes(
-        Request $request,
-        SocieteRepository $societeRepository,
-        ParameterRepository $parameterRepository,
-        EntityManagerInterface $em
-    ) {
-        $onboardingParameter = $parameterRepository->getParameter('bo.onboarding.notification_every', '2 weeks');
-        $form = $this->createForm(ParameterType::class, $onboardingParameter);
-
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em->flush();
-            $this->addFlash('success', 'Interval d\'envoi des emails d\'onboarding mis à jour.');
-
-            return $this->redirectToRoute('app_bo_societes');
-        }
+    public function societes(SocieteRepository $societeRepository) {
 
         return $this->render('bo/societes/societes.html.twig', [
             'societes' => $societeRepository->findAll(),
-            'form' => $form->createView(),
         ]);
     }
 
@@ -67,15 +47,19 @@ class SocieteController extends AbstractController
         Societe $societe,
         Invitator $invitator,
         EntityManagerInterface $em,
-        LicenseService $licenseService
+        LicenseService $licenseService,
+        ProductPrivileges $productPrivileges
     ): Response {
         $admin = $invitator->initUser($societe, RoleSociete::ADMIN);
-        $form = $this->createForm(UserEmailType::class, $admin);
 
-        $form->handleRequest($request);
+        $formUserEmail = $this->createForm(UserEmailType::class, $admin);
+        $formUserEmail->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $invitator->check($admin, $form);
+        $formOnboardingNotificationEvery = $this->createForm(OnboardingNotificationEveryType::class, $societe);
+        $formOnboardingNotificationEvery->handleRequest($request);
+
+        if ($formUserEmail->isSubmitted() && $formUserEmail->isValid()) {
+            $invitator->check($admin, $formUserEmail);
 
             $em->persist($admin);
             $em->flush();
@@ -91,12 +75,56 @@ class SocieteController extends AbstractController
             ]);
         }
 
+        if ($formOnboardingNotificationEvery->isSubmitted() && $formOnboardingNotificationEvery->isValid()) {
+            $em->persist($societe);
+            $em->flush();
+
+            $this->addFlash('success', 'Interval d\'envoi des emails d\'onboarding mis à jour.');
+
+            return $this->redirectToRoute('app_bo_societe', [
+                'id' => $societe->getId(),
+            ]);
+        }
+
         $this->addWarningIfNotInvitationSent($societe);
 
         return $this->render('bo/societes/societe.html.twig', [
             'societe' => $societe,
-            'form' => $form->createView(),
+            'formUserEmail' => $formUserEmail->createView(),
+            'formOnboardingNotificationEvery' => $formOnboardingNotificationEvery->createView(),
             'licenses' => $licenseService->retrieveAllLicenses($societe),
+            'societeProducts' => $productPrivileges->getAllProducts()
+        ]);
+    }
+
+
+    /**
+     * @Route(
+     *     "/societes/product-packs/{id}/{product}",
+     *     name="app_bo_societe_product_packs_switch",
+     *     requirements={"id"="\d+", "product": "^STARTER|STANDARD|PREMIUM$"},
+     *     )
+     */
+    public function societeProductPacks(
+        Societe $societe,
+        string $product,
+        EntityManagerInterface $em,
+        ProductPrivileges $productPrivileges
+    ): Response {
+        $products = $productPrivileges->getAllProducts();
+
+        if (key_exists($product,$products)){
+            $societe->setProductKey($product);
+            $em->persist($societe);
+            $em->flush();
+
+            $this->addFlash('success', "Le pack de fonctionnalités a été mis à jour avec succès.");
+        } else {
+            $this->addFlash('danger', "Une erreur est survenue !!");
+        }
+
+        return $this->redirectToRoute('app_bo_societe', [
+            'id' => $societe->getId(),
         ]);
     }