+++ src/Notification/Sms/RappelSaisieTemps.php
@@ -7,6 +7,8 @@ use App\Entity\SocieteUser;
 use App\Notification\Event\RappelSaisieTempsNotification;
 use App\Repository\SocieteUserRepository;
 use App\Service\DateMonthService;
+use App\SocieteProduct\Product\ProductPrivileges;
+use App\SocieteProduct\ProductPrivilegeCheker;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Twig\Environment as TwigEnvironment;
 
@@ -67,12 +69,14 @@ class RappelSaisieTemps implements EventSubscriberInterface
     public function sendNotificationSaisieTempsAllUsers(RappelSaisieTempsNotification $event): void
     {
         $societe = $event->getSociete();
-        $month = $event->getMonth();
-        $societeUsers = $this->societeUserRepository->findAllNotifiableUsers('notificationSaisieTempsEnabled', $societe);
+        if (ProductPrivilegeCheker::checkProductPrivilege($societe,ProductPrivileges::SMS_NOTIFICATION_SAISIE_TEMPS)){
+            $month = $event->getMonth();
+            $societeUsers = $this->societeUserRepository->findAllNotifiableUsers('notificationSaisieTempsEnabled', $societe);
 
-        foreach ($societeUsers as $societeUser) {
-            if ($societe->getSmsEnabled() && null !== $societeUser->getUser()->getTelephone()) {
-                $this->sendNotificationSaisieTempsSms($societeUser, $month);
+            foreach ($societeUsers as $societeUser) {
+                if ($societe->getSmsEnabled() && null !== $societeUser->getUser()->getTelephone()) {
+                    $this->sendNotificationSaisieTempsSms($societeUser, $month);
+                }
             }
         }
     }